Move Products enum to global interface
