+++ src/lib/interfaces/Global.ts
@@ -7,3 +7,12 @@ export interface ErrorMessage {
     error: string;
     message: string;
 }
+
+export enum Products {
+    idp = "idp",
+    auditor = "auditor",
+    mailer = "mailer",
+    high5 = "high5",
+    fuse = "fuse",
+    all = "all",
+}
